Closer to a successful Windows build.
Nested namespaces are not supported on MSVC. Stop using them. In any
case, namespaces in C++ are a poor substitute for modules and so the
nested namespaces aren't actually that useful.

Use GLEW to help set up OpenGL. This is convenient on Windows to get
access to the latest version of the OpenGL API.

On Windows, boost::filesystem::path.c_str() returns a string with wide
characters. This doesn't work for most APIs that need file names. (e.g
SDL2_image) So, instead call path.string().c_str() to get a const char*
representation of the path.

BUG: EntityX doesn't really support Windows, it turns out. Need to find
a way to get this to link.