Fix holes in the terrain.
The mesher needs a 1 voxel border around the region to be meshed. Create a
VoxelDataStore with at least this border. Then, send the mesher the voxels
it actually needs to extract a surface. Finally, now that the mesher always
gets the voxels it needs and those voxels are always in the valid region of
the VoxelDataStore, we can eliminate the get() method with the default
value fallback.