Bring back OpenGL Command Queue so API is only used on main thread.
Now that we're calling graphics API on background threads for Terrain, we
need to take steps to ensure that OpenGL API is only ever called on the
main thread. The solution is to stick a closure wrapping each OpenGL call
onto a queue and execute later on the main thread.

BUG: Gets a GL_INVALID_OPERATION error pretty consistently when drawing
terrain.