Rolling horizon distance moves back as chunks within it become ready.
We fetch all chunks within this horizon. When chunks within that horizon
are all available then we move the horizon back and fetch some more. This
results in the chunks nearest the camera always being available before
chunks far from the camera.

Fog density is adjusted over time as the horizon distance scrolls back.
This hides the chunks loading at the terrain horizon.