Eliminate SparseArray3D, GridAddressable
Eliminate SparseArray3D because it is not working out. We cannot purge any
items in the LRU list because we do not know whether those items are in use
through references. Likewise, rename BaseArray3D to Array3D, eliminate
SparseVector, and eliminate DenseVector.

Eliminate GridAddressable because it hasn't been particularly valuable and
I intend to create new API in a future commit for accessing a grid by-value
instead of by-reference. Things that did use GridMutable/GridAddressable
now refer to the concrete type Array3D instead.

VoxelDataStore is no longer a ConcurrentGridMutable. The API needed for
VoxelDataStore is going to be slightly different than that used by the
other grids. The only real similarilty is in the transaction-based access
API. In a future commit, I'll remove ConcurrentGridMutable entirely and
replace it with something new. Then, VoxelDataStore will be built on that
something.