Fixed a broken case in MallocZone::reallocate, plus new unit test.
Fixed the case where MallocZone reallocates a block and extends the tail
block backwards into preceding free space.

Also, put a new field in the header which is the offset to the tail block.
This permits the backing memory region to be freed before calling grow().
For example, via realloc().