An instance of TerrainProgressTracker operates with a single chunk.
We pass the tracker around as the chunk is loaded/generated. Progress is
reported to the log and game events are emitted as progress is made.

Next steps: Add a new System to listen to these events and manage a stable
of corresponding wireframe boxes. We'll put wireframe boxes around chunks
in flight, colored to represent the state.