Better hygiene with NSAutoreleasePool when running on macOS.
Also, NSAutoreleasePool documentation states that if we're creating threads
using APIs other than Cocoa (which we are) then we must put Cocoa into
multithreaded mode before using it. We do this by detaching a new NSThread.

In a multithreaded application, Cocoa uses locks and other forms of
internal synchronization to ensure correct behavior at runtime. In the case
of a single threaded application, this synchronization is disabled for
performance reasons.