Implement a fast path for sunlight propagation when fetching for meshes.
When we propagate sunlight across a column, the center column receives
correct light values, and the other columns in the neighborhood receive
partially correct light values. However, there is a region near the center
column which is always correct. So, if the requested region stays within
this safe region then we can avoid propagating sunlight for any column
other than the center column. This reduces the amount of work done.