Change the way the active region is calculated in Terrain.
By itself, batching of meshes provides a modest improvement to terrain
generation time in the scenario, taking 5.8s with a baseline of 6.2s with
no batching.

Combined with a change to the way the active region is calculated, the
terrain generation time in the scenario takes 5.1s. Note that the active
region change by itself gives basically no improvement at all.