Do not call get() in the draw() method. Removes large time spent there.
Add begin()/end() methods to UnlockedSparseGrid so that we can directly
iterate over all elements in the grid. This removes the (substantial) cost
of accessing every grid element via cell-coords each frame.

This also disables frustum culling of terrain chunks in the draw() method.
Though, measurements are indicating that this is counterproductive.